requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --allow-unsafe requirements/prod.in
#
alembic==1.13.1
    # via -r requirements/prod.in
amqp==5.1.1
    # via kombu
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
backports-zoneinfo[tzdata]==0.2.1
    # via
    #   celery
    #   kombu
bcrypt==4.1.2
    # via -r requirements/prod.in
billiard==4.2.0
    # via celery
bleach==6.1.0
    # via -r requirements/prod.in
celery==5.3.6
    # via -r requirements/prod.in
certifi==2023.11.17
    # via
    #   -r requirements/prod.in
    #   sentry-sdk
cffi==1.16.0
    # via
    #   -r requirements/prod.in
    #   cryptography
chameleon==3.8.1
    # via deform
click==8.1.7
    # via
    #   -r requirements/prod.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colander==1.8.2
    # via deform
cryptography==41.0.6
    # via python-jose
data-tasks==0.0.3
    # via -r requirements/prod.in
deform==2.0.15
    # via -r requirements/prod.in
ecdsa==0.18.0
    # via python-jose
elasticsearch==6.8.2
    # via
    #   -r requirements/prod.in
    #   elasticsearch-dsl
elasticsearch-dsl==6.4.0
    # via -r requirements/prod.in
gevent==23.9.1
    # via -r requirements/prod.in
greenlet==2.0.2
    # via
    #   gevent
    #   sqlalchemy
gunicorn==21.2.0
    # via -r requirements/prod.in
h-api==1.1.0
    # via -r requirements/prod.in
h-assets==1.0.6
    # via -r requirements/prod.in
h-matchers==1.2.15
    # via -r requirements/prod.in
h-pyramid-sentry==1.2.4
    # via -r requirements/prod.in
hupper==1.10.2
    # via pyramid
importlib-metadata==4.8.1
    # via
    #   alembic
    #   data-tasks
    #   h-api
    #   h-assets
    #   markdown
importlib-resources==6.1.1
    # via
    #   -r requirements/prod.in
    #   alembic
    #   data-tasks
    #   h-api
    #   jsonschema
    #   jsonschema-specifications
iso8601==0.1.13
    # via
    #   colander
    #   deform
itsdangerous==2.1.2
    # via -r requirements/prod.in
jinja2==2.11.3
    # via
    #   data-tasks
    #   pyramid-jinja2
jsonschema==4.20.0
    # via
    #   -r requirements/prod.in
    #   h-api
jsonschema-specifications==2023.7.1
    # via jsonschema
kombu==5.3.4
    # via
    #   -r requirements/prod.in
    #   celery
mako==1.2.2
    # via alembic
markdown==3.5.1
    # via -r requirements/prod.in
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
    #   pyramid-jinja2
newrelic==9.3.0
    # via -r requirements/prod.in
oauthlib==3.2.2
    # via -r requirements/prod.in
packaging==23.2
    # via
    #   -r requirements/prod.in
    #   gunicorn
    #   zope-sqlalchemy
passlib==1.7.4
    # via -r requirements/prod.in
pastedeploy==2.1.0
    # via plaster-pastedeploy
peppercorn==0.6
    # via deform
pkgutil-resolve-name==1.3.10
    # via jsonschema
plaster==1.0
    # via
    #   plaster-pastedeploy
    #   pyramid
plaster-pastedeploy==0.7
    # via pyramid
prompt-toolkit==3.0.18
    # via click-repl
psycogreen==1.0.2
    # via -r requirements/prod.in
psycopg2==2.9.9
    # via
    #   -r requirements/prod.in
    #   data-tasks
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pycryptodomex==3.19.0
    # via -r requirements/prod.in
pyjwt==2.8.0
    # via -r requirements/prod.in
pyparsing==3.1.1
    # via -r requirements/prod.in
pyramid==2.0.2
    # via
    #   -r requirements/prod.in
    #   h-assets
    #   h-pyramid-sentry
    #   pyramid-exclog
    #   pyramid-jinja2
    #   pyramid-mailer
    #   pyramid-retry
    #   pyramid-sanity
    #   pyramid-services
    #   pyramid-tm
pyramid-exclog==1.1
    # via -r requirements/prod.in
pyramid-jinja2==2.10
    # via -r requirements/prod.in
pyramid-mailer==0.15.1
    # via -r requirements/prod.in
pyramid-retry==2.1.1
    # via -r requirements/prod.in
pyramid-sanity==1.0.3
    # via -r requirements/prod.in
pyramid-services==2.2
    # via -r requirements/prod.in
pyramid-tm==2.5
    # via -r requirements/prod.in
python-dateutil==2.8.2
    # via
    #   -r requirements/prod.in
    #   celery
    #   elasticsearch-dsl
python-jose[cryptography]==3.3.0
    # via -r requirements/prod.in
python-slugify==8.0.1
    # via -r requirements/prod.in
pytz==2023.3.post1
    # via -r requirements/prod.in
referencing==0.30.2
    # via
    #   h-api
    #   jsonschema
    #   jsonschema-specifications
repoze-sendmail==4.3
    # via pyramid-mailer
rfc3339-validator==0.1.4
    # via -r requirements/prod.in
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via python-jose
sentry-sdk==1.17.0
    # via h-pyramid-sentry
six==1.15.0
    # via
    #   bleach
    #   click-repl
    #   ecdsa
    #   elasticsearch-dsl
    #   python-dateutil
    #   rfc3339-validator
sqlalchemy==1.4.50
    # via
    #   -r requirements/prod.in
    #   alembic
    #   data-tasks
    #   zope-sqlalchemy
sqlparse==0.4.4
    # via
    #   -r requirements/prod.in
    #   data-tasks
supervisor==4.2.5
    # via -r requirements/prod.in
tabulate==0.9.0
    # via
    #   -r requirements/prod.in
    #   data-tasks
text-unidecode==1.3
    # via python-slugify
transaction==4.0
    # via
    #   -r requirements/prod.in
    #   pyramid-mailer
    #   pyramid-tm
    #   repoze-sendmail
    #   zope-sqlalchemy
translationstring==1.4
    # via
    #   colander
    #   deform
    #   pyramid
typing-extensions==4.5.0
    # via
    #   alembic
    #   kombu
tzdata==2023.3
    # via
    #   backports-zoneinfo
    #   celery
urllib3==1.26.18
    # via
    #   elasticsearch
    #   sentry-sdk
venusian==3.1.0
    # via
    #   -r requirements/prod.in
    #   pyramid
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
webob==1.8.7
    # via pyramid
wired==0.2.2
    # via pyramid-services
ws4py==0.5.1
    # via -r requirements/prod.in
wsaccel==0.6.6
    # via -r requirements/prod.in
zipp==3.4.1
    # via
    #   importlib-metadata
    #   importlib-resources
zope-deprecation==4.1.2
    # via
    #   deform
    #   pyramid
    #   pyramid-jinja2
zope-event==4.4
    # via gevent
zope-interface==6.1
    # via
    #   -r requirements/prod.in
    #   gevent
    #   pyramid
    #   pyramid-retry
    #   pyramid-services
    #   repoze-sendmail
    #   transaction
    #   wired
    #   zope-sqlalchemy
zope-sqlalchemy==3.1
    # via -r requirements/prod.in

# The following packages are considered to be unsafe in a requirements file:
setuptools==69.0.2
    # via
    #   colander
    #   plaster
    #   pyramid
    #   repoze-sendmail
    #   supervisor
    #   zope-deprecation
    #   zope-event
    #   zope-interface
    #   zope-sqlalchemy
